refactor: move locale handling from root layout
