+++ apps/web/app/[locale]/layout.tsx
@@ -1,0 +1,30 @@
+import { NextIntlClientProvider } from 'next-intl';
+import { notFound } from 'next/navigation';
+import { locales } from '@/utils/locales';
+
+export function generateStaticParams() {
+  return locales.map((locale) => ({ locale }));
+}
+
+export default async function LocaleLayout({
+  children,
+  params: { locale },
+}: {
+  children: React.ReactNode;
+  params: { locale: string };
+}) {
+  if (!locales.includes(locale as typeof locales[number])) {
+    notFound();
+  }
+
+  const messages = {
+    common: (await import(`@/locales/${locale}/common.json`)).default,
+  };
+
+  return (
+    <NextIntlClientProvider locale={locale} messages={messages}>
+      {children}
+    </NextIntlClientProvider>
+  );
+}
+
